Graded HW5
Docstrings are present for functions, but are not detailed enough or describe self-commenting code. Be sure to explain where numbers are coming if they don't obviously come from earlier in the code. Nice use of in line comments at points **Documentation: +1**

There are unnecessary loops and the controller is overly complex. Move function can be condensed into one function. Variable names are clear and the code is easy to follow though! Define Classes in the order they are used. **Style: +0.75**

Code runs as intended. Game is super cute and fun to play, though usability is a concern because of the vague hit-box. **Functionality: +3**

Nicely done! :)
Grade: **4.75/5**